Move CommaSeparatedList into a different module
It’s a generic util field
